refactor : 추후 새로운 책 데이터소스를 추가할 수 있도록 리팩토링
