+++ README.md
@@ -6,7 +6,7 @@
 - 북마크 기능 (추가, 삭제)
 - 회원가입 기능 (패스워드 암호화)
 - 책 상세 정보 조회
-- 비지니스 로직 위주로 테스트 케이스 작성
+- 추후 새로운 책 데이터 소스(새로운 책 정보 API)를 추가 할 수 있도록 확장성을 고려하여 개발
 
 ## 사용한 라이브러리 및 오픈소스
 - Java8
