+++ src/main/java/com/ooa1769/bs/book/support/BookService.java
@@ -1,9 +1,6 @@
 package com.ooa1769.bs.book.support;
 
-import com.ooa1769.bs.book.Book;
-import com.ooa1769.bs.book.BookMark;
-import com.ooa1769.bs.book.BookNotFoundException;
-import com.ooa1769.bs.book.SearchOption;
+import com.ooa1769.bs.book.*;
 import com.ooa1769.bs.member.Member;
 import com.ooa1769.bs.member.support.MemberRepository;
 import com.ooa1769.bs.web.dto.BookMarkDto;
@@ -26,12 +23,12 @@ public class BookService {
         this.bookMarkRepository = bookMarkRepository;
     }
 
-    public Page<Book> getBooksByKeyword(SearchOption searchOption) {
-        return searchService.search(searchOption);
+    public Page<Book> getBooksByKeyword(ApiSearchOption apiSearchOption) {
+        return searchService.search(apiSearchOption);
     }
 
-    public Book getBookByIsbn(SearchOption searchOption) {
-        Page<Book> pageBook = searchService.search(searchOption);
+    public Book getBookByIsbn(ApiSearchOption apiSearchOption) {
+        Page<Book> pageBook = searchService.search(apiSearchOption);
         if (pageBook.getTotalElements() == 0) {
             throw new BookNotFoundException("해당 책이 존재하지 않습니다.");
         }
