+++ src/main/java/com/ooa1769/bs/config/SecurityConfig.java
@@ -34,6 +34,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         http
                 .authorizeRequests()
                 .antMatchers("/","/css/**","/js/**","/favicon.ico"
+                        ,"/404", "/500"
                         ,"/h2-console/**", "/member/form", "/member/register", "/member/congratulation")
                 .permitAll()
                 .anyRequest().authenticated()
