+++ src/main/java/com/ooa1769/bs/config/WebMvcConfig.java
@@ -1,9 +1,18 @@
 package com.ooa1769.bs.config;
 
+import com.ooa1769.bs.support.domain.ApiTypeConverter;
 import com.ooa1769.bs.support.security.LoginUserHandlerMethodArgumentResolver;
+import org.springframework.boot.context.config.ResourceNotFoundException;
+import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
+import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
+import org.springframework.boot.web.servlet.ErrorPage;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
+import org.springframework.format.FormatterRegistry;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
@@ -25,6 +34,8 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter {
     public void addViewControllers(ViewControllerRegistry registry) {
         registry.addViewController("/").setViewName("home");
         registry.addViewController("/login").setViewName("loginPage");
+        registry.addViewController("/404").setViewName("error/404");
+        registry.addViewController("/500").setViewName("error/500");
         registry.setOrder(Ordered.HIGHEST_PRECEDENCE);
     }
 
@@ -36,5 +47,25 @@ public class WebMvcConfig extends WebMvcConfigurerAdapter {
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
         argumentResolvers.add(loginUserArgumentResolver());
+    }
+
+    @Override
+    public void addFormatters(FormatterRegistry registry){
+        registry.addConverter(new ApiTypeConverter());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    public void handleException(ResourceNotFoundException ex) {
+    }
+
+    @Configuration
+    public static class ErrorConfig implements EmbeddedServletContainerCustomizer {
+
+        @Override
+        public void customize(ConfigurableEmbeddedServletContainer factory) {
+            factory.addErrorPages(new ErrorPage(HttpStatus.NOT_FOUND, "/404"));
+            factory.addErrorPages(new ErrorPage(HttpStatus.INTERNAL_SERVER_ERROR, "/500"));
+        }
     }
 }
