+++ src/main/java/com/ooa1769/bs/support/domain/ApiTypeConverter.java
@@ -1,0 +1,11 @@
+package com.ooa1769.bs.support.domain;
+
+import com.ooa1769.bs.book.ApiType;
+import org.springframework.core.convert.converter.Converter;
+
+public class ApiTypeConverter implements Converter<String, ApiType> {
+
+    public ApiType convert(String value) {
+        return ApiType.code(value);
+    }
+}
