+++ src/main/java/com/ooa1769/bs/web/api/ApiBookController.java
@@ -1,10 +1,9 @@
 package com.ooa1769.bs.web.api;
 
-import com.ooa1769.bs.book.SearchOption;
+import com.ooa1769.bs.book.support.ApiSearchOption;
 import com.ooa1769.bs.book.support.BookService;
 import com.ooa1769.bs.book.support.SearchHistoryService;
 import com.ooa1769.bs.member.Member;
-import com.ooa1769.bs.member.support.MemberService;
 import com.ooa1769.bs.support.security.LoginMember;
 import com.ooa1769.bs.support.util.Mappings;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,13 +27,8 @@ public class ApiBookController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<?> search(@LoginMember Member member, SearchOption searchOption) {
-        searchHistoryService.addSearchHistory(member.getEmail(), searchOption.getQuery());
-        return new ResponseEntity<>(bookService.getBooksByKeyword(searchOption), HttpStatus.OK);
+    public ResponseEntity<?> search(@LoginMember Member member, ApiSearchOption apiSearchOption) {
+        searchHistoryService.addSearchHistory(member, apiSearchOption.getQuery());
+        return new ResponseEntity<>(bookService.getBooksByKeyword(apiSearchOption), HttpStatus.OK);
     }
-
-/*    @GetMapping("/keywords")
-    public ResponseEntity<?> getKeywords(@LoginMember Member member) {
-        return new ResponseEntity<>(, HttpStatus.OK);
-    }*/
 }
