+++ src/main/resources/templates/book/view.html
@@ -47,7 +47,8 @@
             <p><a th:href="@{ /book(page=${searchOption.page},
 			                size=${searchOption.size},
 			                query=${keyword},
-			                target=${type}
+			                target=${type},
+			                apiType=${searchOption.apiType.code}
 			             )}" class="btn btn-small btn-primary pull-right" >목록으로</a>
             </p>
         </div>
