+++ src/main/resources/templates/error/404.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/web/thymeleaf/layout"
+      layout:decorator="layout">
+<head>
+    <title>404에러</title>
+</head>
+<body>
+<h1 layout:fragment="header">404에러</h1>
+<div layout:fragment="content" class="container">
+
+</div>
+</body>
+</html>
